+++ pyPhenology/models/__init__.py
@@ -1,5 +1,8 @@
-from .base import ThermalTime, Uniforc, Unichill, Alternating, \
-                  MSB, BootstrapModel, Linear
+from .thermaltime import ThermalTime
+from .chuine_models import Uniforc, Unichill
+from .alternating import Alternating, MSB
+from .ensemble_models import BootstrapModel
+from .stat_models import Linear
 
 __all__ = ['ThermalTime',
            'Uniforc',
