+++ pyPhenology/models/base.py
@@ -3,7 +3,6 @@ import pandas as pd
 from . import utils, validation
 import time
 from collections import OrderedDict
-from copy import deepcopy
 
 class _base_model():
     def __init__(self):
@@ -267,367 +266,3 @@ class _base_model():
     
     def score(self, metric='rmse'):
         raise NotImplementedError()
-
-
-
-
-
-class Alternating(_base_model):
-    """Alternating model, originally defined in Cannell & Smith 1983.
-    Phenological event happens the first day that forcing is greater 
-    than an exponential curve of number of chill days.
-    
-    Parameters
-    ----------
-    a : int | float
-        Intercept of chill day curve
-    
-    b : int | float
-        Slope of chill day curve
-    
-    c : int | float
-        scale parameter of chill day curve
-        
-    threshold : int | flaot
-        Degree threshold above which forcing accumulates, and
-        below which chilling accumulates. Set to 5 (assuming C)
-        by default.
-        
-    t1 : int
-        DOY which forcing and chilling accumulationg starts. Set
-        to 1 (Jan 1) by default.
-    """
-    def __init__(self, parameters={}):
-        _base_model.__init__(self)
-        self.all_required_parameters = {'a':(-1000,1000), 'b':(0,5000), 'c':(-5,0),
-                                        'threshold':(5,5), 't1':(1,1)}
-        self._organize_parameters(parameters)
-    
-    def _apply_model(self, temperature, doy_series, a, b, c, threshold, t1):
-        chill_days = ((temperature < threshold)*1).copy()
-        chill_days[doy_series < t1]=0
-        chill_days = utils.forcing_accumulator(chill_days)
-
-        # Accumulated growing degree days from Jan 1
-        gdd = temperature.copy()
-        gdd[gdd < threshold]=0
-        gdd[doy_series < t1]=0
-        gdd = utils.forcing_accumulator(gdd)
-
-        # Phenological event happens the first day gdd is > chill_day curve
-        chill_day_curve = a + b * np.exp( c * chill_days)
-        difference = gdd - chill_day_curve
-
-        # The estimate is equal to the first day that
-        # gdd - chill_day_curve > 0
-        return utils.doy_estimator(difference, doy_series, threshold=0)
-
-
-class BootstrapModel():
-    """Fit a model using bootstrapping of the data.
-
-    """
-    def __init__(self, core_model, num_bootstraps, parameters={}):
-        validation.validate_model(core_model())
-        
-        self.model_list=[]
-        if isinstance(parameters, str):
-            # A filename pointing toward a file from save_params()
-            params = pd.read_csv(parameters).to_dict('records')
-            for bootstrap_iteration in params:
-                bootstrap_iteration.pop('bootstrap_num')
-                self.model_list.append(core_model(parameters=bootstrap_iteration))
-        elif isinstance(parameters, dict):
-            for i in range(num_bootstraps):            
-                self.model_list.append(core_model(parameters=parameters))
-        elif isinstance(parameters, list):
-            # If its the output of BootstrapModel.get_params()
-            for bootstrap_iteration in parameters:
-                bootstrap_iteration.pop('bootstrap_num')
-                self.model_list.append(core_model(parameters=bootstrap_iteration))
-        else:
-            raise TypeError('parameters must be str or dict, got: ' + str(type(parameters)))
-    
-    def fit(self,observations, temperature, **kwargs):
-        #TODO: do the temperature transform here cause so it doesn't get reapated a bunch
-        # need to wait till fit takes arrays directly
-        validation.validate_observations(observations)
-        validation.validate_temperature(temperature)
-        for model in self.model_list:
-            obs_shuffled = observations.sample(frac=1, replace=True).copy()
-            model.fit(obs_shuffled, temperature, **kwargs)
-
-    def predict(self,to_predict=None, temperature=None, aggregation='mean', **kwargs):
-        """Make predictions from the bootstrapped models.
-        Predictions will be made using each of the bootstrapped models, with
-        the final results being the mean or median (or other) of all bootstraps.
-        
-        Parameters
-        ----------
-        aggregation : str
-            Either 'mean','median', or 'none'. 'none' return *all* predictions
-            in an array of size (num_bootstraps, num_samples)
-        
-        """
-        #TODO: do the temperature transform here cause so it doesn't get reapated a bunch
-        # need to wait till predict takes arrays directly
-        predictions=[]
-        for model in self.model_list:
-            predictions.append(model.predict(to_predict=to_predict, 
-                                             temperature=temperature,
-                                             **kwargs))
-        
-        predictions = np.array(predictions)
-        if aggregation=='mean':
-            predictions = np.mean(predictions, 0)
-        elif aggregation=='median':
-            predictions = np.median(predictions, 0)
-        elif aggregation=='none':
-            pass
-        else:
-            raise ValueError('Unknown aggregation: ' + str(aggregation))
-        
-        return predictions
-
-    def get_params(self):
-        all_params=[]
-        for i, model in enumerate(self.model_list):
-            all_params.append(deepcopy(model.get_params()))
-            all_params[-1].update({'bootstrap_num':i})
-
-        return all_params
-
-    def save_params(self, filename):
-        assert len(self.model_list[0]._fitted_params)>0, 'Parameters not fit, nothing to save'
-        params = self.get_params()
-        pd.DataFrame(params).to_csv(filename, index=False)
-
-class ThermalTime(_base_model):
-    """The classic growing degree day model using
-    a fixed threshold above which forcing accumulates.
-    
-    Parameters
-    ----------
-    t1 : int
-        The DOY which forcing accumulating beings
-    
-    T : int
-        The threshold above which forcing accumulates
-    
-    F : int, > 0
-        The total forcing units required
-    """
-    def __init__(self, parameters={}):
-        _base_model.__init__(self)
-        self.all_required_parameters = {'t1':(-67,298),'T':(-25,25),'F':(0,1000)}
-        self._organize_parameters(parameters)
-    
-    def _apply_model(self, temperature, doy_series, t1, T, F):
-        #Temperature threshold
-        temperature[temperature<T]=0
-    
-        #Only accumulate forcing after t1
-        temperature[doy_series<t1]=0
-    
-        accumulated_gdd=utils.forcing_accumulator(temperature)
-    
-        return utils.doy_estimator(forcing = accumulated_gdd, 
-                                   doy_series = doy_series, 
-                                   threshold = F)
-
-class Uniforc(_base_model):
-    """Single phase forcing model using a 
-    sigmoid function for forcing units.
-    Chuine 2000
-    
-    Parameters
-    ----------
-    t1 : int
-        The DOY which forcing accumulating beings
-    
-    F : int, > 0
-        The total forcing units required
-        
-    b : int
-        Sigmoid function parameter
-    
-    c : int
-        Sigmoid function parameter
-    """
-    def __init__(self, parameters={} ):
-        _base_model.__init__(self)
-        self.all_required_parameters = {'t1':(-67,298),'F':(0,200),'b':(-20,0),'c':(-50,50)}
-        self._organize_parameters(parameters)
-    
-    def _apply_model(self, temperature, doy_series, t1, F, b, c):
-        temperature = utils.sigmoid2(temperature, b=b, c=c)
-    
-        #Only accumulate forcing after t1
-        temperature[doy_series<t1]=0
-    
-        accumulateed_forcing=utils.forcing_accumulator(temperature)
-    
-        return utils.doy_estimator(forcing = accumulateed_forcing,
-                                   doy_series=doy_series,
-                                   threshold=F)
-
-class Unichill(_base_model):
-    """Two phase forcing model using a 
-    sigmoid function for forcing units 
-    and chilling. 
-    Chuine 2000
-    
-    Parameters
-    ----------
-    t0 : int
-        The DOY which chilling accumulating beings
-    
-    C : int, > 0
-        The total chilling units required
-
-    F : int, > 0
-        The total forcing units required
-        
-    b_f : int
-        Sigmoid function parameter for forcing
-    
-    c_f : int
-        Sigmoid function parameter for forcing
-        
-    a_c : int
-        Sigmoid funcion parameter for chilling
-        
-    b_c : int
-        Sigmoid funcion parameter for chilling
-        
-    c_c : int
-        Sigmoid funcion parameter for chilling
-    """
-    def __init__(self, parameters={}):
-        _base_model.__init__(self)
-        self.all_required_parameters = {'t0':(-67,298),'C':(0,300),'F':(0,200),
-                                        'b_f':(-20,0),'c_f':(-50,50),
-                                        'a_c':(0,20),'b_c':(-20,20),'c_c':(-50,50)}
-        self._organize_parameters(parameters)
-    
-    def _apply_model(self, temperature, doy_series, t0, C, F, b_f, c_f, a_c, b_c, c_c):
-        assert len(temperature.shape)==2, 'Unichill model currently only supports 2d temperature arrays'
-
-        temp_chilling = temperature.copy()
-        temp_forcing  = temperature.copy()
-        
-        temp_forcing = utils.sigmoid2(temp_forcing, b=b_f, c=c_f)
-        temp_chilling =utils.sigmoid3(temp_chilling, a=a_c, b=b_c, c=c_c) 
-    
-        #Only accumulate chilling after t0
-        temp_chilling[doy_series<t0]=0
-        accumulated_chill=utils.forcing_accumulator(temp_chilling)
-        
-        # Heat forcing accumulation starts when the chilling
-        # requirement, C, has been met(t1 in the equation). 
-        # Enforce this by setting everything prior to that date to 0
-        # TODO: optimize this so it doesn't use a for loop
-        t1_values = utils.doy_estimator(forcing = accumulated_chill,
-                                        doy_series=doy_series,
-                                        threshold=C)
-        for col, t1 in enumerate(t1_values):
-            temp_forcing[doy_series<t1, col]=0
-    
-        accumulated_forcing = utils.forcing_accumulator(temp_forcing)
-        
-        return utils.doy_estimator(forcing = accumulated_forcing,
-                                   doy_series=doy_series,
-                                   threshold=F)
-
-class MSB(_base_model):
-    """Macroscale Species-specific Budburst model. Jeong et al. 2013
-    Extension of the Alternating model which add a correction (d)
-    using the mean spring temperature 
-    
-    Parameters
-    ----------
-    a : int | float
-        Intercept of chill day curve
-    
-    b : int | float
-        Slope of chill day curve
-    
-    c : int | float
-        scale parameter of chill day curve
-    
-    d : int | float
-        Correction factor
-        
-    threshold : int | flaot
-        Degree threshold above which forcing accumulates, and
-        below which chilling accumulates. Set to 5 (assuming C)
-        by default.
-        
-    t1 : int
-        DOY which forcing and chilling accumulationg starts. Set
-        to 1 (Jan 1) by default.
-    """
-    def __init__(self, parameters={}):
-        _base_model.__init__(self)
-        self.all_required_parameters = {'a':(-1000,1000), 'b':(0,5000), 'c':(-5,0),
-                                        'd':(-100,100), 'threshold':(5,5), 't1':(1,1)}
-        self._organize_parameters(parameters)
-    
-    def _apply_model(self, temperature, doy_series, a, b, c, d, threshold, t1):
-        chill_days = ((temperature < threshold)*1).copy()
-        chill_days[doy_series < t1]=0
-        chill_days = utils.forcing_accumulator(chill_days)
-
-        # Accumulated growing degree days from Jan 1
-        gdd = temperature.copy()
-        gdd[gdd < threshold]=0
-        gdd[doy_series < t1]=0
-        gdd = utils.forcing_accumulator(gdd)
-
-        chill_day_curve = a + b * np.exp( c * chill_days)
-        
-        # Make the spring temps the same shape as chill_day_curve
-        # for easy addition.
-        mean_spring_temp = utils.mean_temperature(temperature, doy_series,
-                                                  start_doy=0, end_doy=60)
-        mean_spring_temp *= d
-        # Add in correction based on per site spring temperature
-        chill_day_curve += mean_spring_temp
-
-        # Phenological event happens the first day gdd is > chill_day curve
-        difference = gdd - chill_day_curve
-
-        # The estimate is equal to the first day that
-        # gdd - chill_day_curve > 0
-        return utils.doy_estimator(difference, doy_series, threshold=0)
-
-class Linear(_base_model):
-    """A linear regression where DOY ~ mean_spring_tempearture
-    
-    Parameters
-    ----------
-    intercept : int | float
-        y intercept of the model
-    
-    slope : int | float
-        Slope of the model
-        
-    spring_start : int
-        The start day of spring, defaults to Jan 1 (DOY 0)
-    
-    spring_end : int
-        The last day of spring, defaults to March 30 (DOY 90)
-
-    """
-    def __init__(self, parameters={}):
-        _base_model.__init__(self)
-        self.all_required_parameters = {'intercept':(-67,298),'slope':(-25,25),
-                                        'spring_start':(0,0), 'spring_end':(90,90)}
-        self._organize_parameters(parameters)
-    
-    def _apply_model(self, temperature, doy_series, intercept, slope, 
-                     spring_start, spring_end):
-        mean_spring_temp = utils.mean_temperature(temperature, doy_series,
-                                                  start_doy = spring_start,
-                                                  end_doy = spring_end)
-        return mean_spring_temp * slope + intercept
